+++ src/pages/main/MainPage.tsx
@@ -1,35 +1,46 @@
-import { CurrentWeather } from './components/CurrentWeather';
-import { HourlyWeather } from './components/HourlyWeather';
-import { TodayOutfitRecommendation } from './components/TodayOutfitRecommendation';
+// src/pages/MainPage.tsx
+import { Box } from '@mui/material';
 import {
   ComponentsGrid,
   FullWidthCard,
   MainContainer,
   PlaceholderCard,
-  RecommendCard,
 } from './styles/MainPageContentStyles';
+import { CurrentWeather } from './components/CurrentWeather';
+import { TodayOutfitRecommendation } from './components/TodayOutfitRecommendation';
+import { HourlyWeather } from './components/HourlyWeather';
 
-interface WeatherData {
-  location: string;
-  temperature: number;
-  condition: string;
-}
-export const Mainpage = () => {
+export const MainPage = () => {
   return (
     <MainContainer>
       <ComponentsGrid>
         {/* 현재 날씨 */}
-        <CurrentWeather />
+        <CurrentWeather
+          location='서울'
+          temperature={18}
+          condition='scattered clouds'
+          iconCode='03d'
+          precipitation={30}
+          feelsLike={16}
+          onEditLocation={() => {
+            console.log('위치 변경');
+          }}
+        />
 
+        {/* 오늘의 추천 코디 */}
         <TodayOutfitRecommendation temperature={18} />
 
-        <HourlyWeather />
+        {/* 시간별 날씨 */}
+        <HourlyWeather hourlyData={[]} />
 
         <FullWidthCard>AI 추천 문구</FullWidthCard>
 
-        <RecommendCard>즐겨 찾는 지역</RecommendCard>
+        <PlaceholderCard>즐겨 찾는 지역</PlaceholderCard>
 
-        <RecommendCard>즐겨 찾는 지역 의상 추천</RecommendCard>
+        <PlaceholderCard>○○○경기장</PlaceholderCard>
+
+        {/* 하단 추가 정보 */}
+        <FullWidthCard>지금, ○○○경기장에 가시는건가요?</FullWidthCard>
       </ComponentsGrid>
     </MainContainer>
   );
